requirements-full.txt
#ADK and Gemini
google-adk==1.23.0
google-genai==1.60.0
google-cloud-aiplatform==1.134.0
google-auth==2.47.0

#Sigma (pysigma-backend-elasticsearch requires pysigma>=0.11.3)
pysigma>=0.11.3,<0.12.0
pysigma-backend-elasticsearch==1.1.0
pysigma-pipeline-sysmon==1.0.0

#data validation
pydantic==2.12.5
pydantic-settings==2.12.0
PyYAML==6.0.3

#testing
elasticsearch==8.12.0
docker==7.0.0
pytest==8.0.0

#utilities
python-dotenv==1.2.1
tenacity==9.1.2
